requirements.txt
asgiref==3.11.0
boto3==1.41.5
botocore==1.41.5
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==3.0.1
Django==4.2.25
django-cors-headers==4.3.1
django-extensions==4.1
django-filter==24.2
django-storages==1.14.6
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
djoser==2.2.2
fonttools==4.60.1
gunicorn==22.0.0
html5lib==1.1
idna==3.11
jmespath==1.0.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.9
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.15.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.4
tinycss2==1.4.0
typing_extensions==4.15.0
urllib3==1.26.20
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3.post1
